remove id from logoComponent, replace boolean props to color (logo-component is more flexible now), move footer from app.js to new component
