+++ src/components/icons/webpurple-logo-icon.jsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const WebpurpleLogoIcon = ({ isGray }) => (
+const WebpurpleLogoIcon = ({ color }) => (
     <svg
         width="35px"
         height="31px"
@@ -29,20 +29,20 @@ const WebpurpleLogoIcon = ({ isGray }) => (
                 <g transform="translate(516.000000, 169.000000)">
                     <g transform="translate(96.000000, 40.000000)">
                         <g>
-                            <g transform="translate(-0.000000, 0.206061)" id="logo-icon">
+                            <g transform="translate(-0.000000, 0.206061)">
                                 <polygon
-                                    fill={isGray ? '#b3a8c1' : 'url(#linearGradient-1)'}
+                                    fill={color || 'url(#linearGradient-1)'}
                                     points="25.884174 30.7265263 25.9148433 0.0504544753 34.522761 5.15369572 34.4920916 25.6232851" />
                                 <polygon
-                                    fill={isGray ? '#b3a8c1' : 'url(#linearGradient-2)'}
+                                    fill={color || 'url(#linearGradient-2)'}
                                     points="0.0543461312 25.6232851 8.66226382 30.7265263 17.2762563 25.6232851 25.884174 30.7265263 34.4920916 25.6232851 17.2762563 15.4107988" />
                                 <polygon
-                                    fill={isGray ? '#b3a8c1' : 'url(#linearGradient-3)'}
-                                    opacity={isGray ? 1 : 0.4}
-                                    style={{ mixBlendMode: isGray ? 'normal' : 'multiply' }}
+                                    fill={color || 'url(#linearGradient-3)'}
+                                    opacity={color ? 1 : 0.4}
+                                    style={{ mixBlendMode: color ? 'normal' : 'multiply' }}
                                     points="8.66226382 30.7265263 17.2762563 15.4107988 0.0543461312 25.6232851" />
                                 <polygon
-                                    fill={isGray ? '#b3a8c1' : 'url(#linearGradient-4)'}
+                                    fill={color || 'url(#linearGradient-4)'}
                                     points="0.0543461312 25.6232851 0.150496129 5.15369624 8.75841382 0.0504549893 8.66226382 30.7265263" />
                             </g>
                         </g>
