+++ src/containers/app/app.jsx
@@ -7,7 +7,7 @@ import { renderRoutes } from 'react-router-config';
 import { media } from '../../utils/css-utils';
 
 import LoginPopup from '../../components/login/login-popup';
-import WebpurpleLogoIcon from '../../components/icons/webpurple-logo-icon';
+import FooterComponent from '../../components/footer';
 
 const Container = styled.div`
     margin: 0 auto;
@@ -63,65 +63,6 @@ const NavigationLink = styled(NavLink)`
     }
 `;
 
-const Footer = styled.footer`
-    bottom: 0;
-    background-color: ${props => props.theme.grape};
-    box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.3);
-    padding: 3rem 2rem 3.5rem;
-    ${media.tablet`
-        display: flex;
-        padding: 4rem 2rem 3rem;
-        align-items: center;
-        justify-content: space-between;
-    `}
-    ${media.desktop`padding: 4rem 10rem 3rem;`}
-`;
-
-const FooterElementsStyle = styled.p`
-    margin: 0;
-    color: #fff;
-    font-family: 'Oxygen', sans-serif;
-    font-size: 1.6rem;
-    font-weight: bold;
-    text-align: center;
-    ${media.phone`
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        text-align: left;
-    `}
-    ${media.tablet`font-size: 1.8rem;`}
-`;
-
-const BlockElem = styled.p`
-    margin: 1rem 0;
-    text-align: center;
-    ${media.phone`text-align: left;`}
-`;
-
-const LogoContainer = styled(BlockElem)`
-    ${media.phone`padding-right: 0.8rem;`}
-    
-`;
-
-const Copyrights = styled(FooterElementsStyle)`
-    opacity: 0.6;
-    ${media.tablet`margin-right: auto;`}
-`;
-
-const Contacts = styled(FooterElementsStyle)`
-    ${media.tablet`margin-left: auto;`}
-`;
-
-const MailInfo = styled.a`
-    color: ${props => props.theme.lipstick};
-    text-decoration: none;
-    
-    &:hover {
-        text-decoration: underline;
-    }
-`;
-
 const theme = {
     grape: '#432867',
     warmGrey: '#a1a1a1',
@@ -150,14 +91,7 @@ const AppContainer = ({ route }) => (
             <main>
                 {renderRoutes(route.routes)}
             </main>
-            <Footer>
-                <Copyrights>
-                    <LogoContainer>
-                        <WebpurpleLogoIcon isGray />
-                    </LogoContainer><BlockElem>&#169; 2013 - {new Date().getFullYear()} Webpurple.</BlockElem><BlockElem>All rights reserved.</BlockElem>
-                </Copyrights>
-                <Contacts>Contact us <MailInfo href="mailto:info@webpurle.com">info@webpurle.com</MailInfo></Contacts>
-            </Footer>
+            <FooterComponent />
         </Container>
     </ThemeProvider>
 );
